light-archive-mcp/requirements_uv.txt
# Light Archive MCP Server Dependencies (uv용)
# uv가 자동으로 처리하므로 mcp는 제외

# Data validation
pydantic>=2.0.0

# HTTP client
httpx>=0.25.0

# Supabase client
supabase>=2.0.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Image processing (NO Base64 - direct binary)
Pillow>=10.0.0
